Move utils module to src dir
